requirements-dev.txt
django<2.1
djangorestframework
